[UPDATE] 파일 삭제 확인 추가 #84

+++ src/components/files/FilePreviewer.jsx
@@ -28,16 +28,19 @@ const FilePreviewer = ({ file, ...rest}) => {
 
 		const s3 = new AWS.S3();
 
-		s3.deleteObject(params, (e, data) => {
-			// 삭제 중 에러 발생 시
-			if (e) {
-				alert("파일 삭제 중 오류가 발생했습니다:", e.message, e.stack);
-				return;
-			};
-			alert("파일 삭제에 성공했습니다.");
-			dispatch(fileActions.deletePreview(file.awsFileName))
-		});
-
+		const result = window.confirm("파일을 정말로 삭제하시겠습니까?");
+    if (result) {
+      s3.deleteObject(params, (e, data) => {
+				// 삭제 중 에러 발생 시
+				if (e) {
+					alert("파일 삭제 중 오류가 발생했습니다:", e.message, e.stack);
+					return;
+				};
+				alert("파일 삭제에 성공했습니다.");
+				dispatch(fileActions.deletePreview(file.awsFileName))
+			});	
+    } else return;
+		
 		// const deleteFile = remove.promise()	
 		// deleteFile.then(
 		// 	function (data) {
